+++ src/math/sphere_intersect.c
@@ -17,16 +17,16 @@
 t_vec3	sphere_intersect(t_sphere *sphere, t_vec3 pos, t_vec3 dir)
 {
 	t_vec3 oc = vec3_sub(pos, sphere->pos);
-	float a = vec3_dot(dir, dir);
-	float b = 2 * vec3_dot(oc, dir);
-	float c = vec3_dot(oc, oc) - sphere->radius * sphere->radius;
-	float discriminant = b * b - 4 * a * c;
+	double a = vec3_dot(dir, dir);
+	double b = 2 * vec3_dot(oc, dir);
+	double c = vec3_dot(oc, oc) - sphere->radius * sphere->radius;
+	double discriminant = b * b - 4 * a * c;
 
 	if (discriminant < 0)
 		return (vec3_init(NAN, NAN, NAN));
-	float t = (-b - sqrtf(discriminant)) / (2 * a);
+	double t = (-b - sqrt(discriminant)) / (2 * a);
 	if (t < 0)
-		t = (-b + sqrtf(discriminant)) / (2 * a);
+		t = (-b + sqrt(discriminant)) / (2 * a);
 	if (t < 0)
 		return (vec3_init(NAN, NAN, NAN));
 	return (vec3_add(pos, vec3_mul(dir, t)));
