+++ src/math/vector.c
@@ -6,7 +6,7 @@
 /*   By: kschmidt <kevin@imkx.dev>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 21:22:52 by kschmidt          #+#    #+#             */
-/*   Updated: 2023/05/02 21:24:10 by kschmidt         ###   ########.fr       */
+/*   Updated: 2023/05/04 13:57:22 by kschmidt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ t_vec3	vec3_sub(t_vec3 a, t_vec3 b)
 	return ((t_vec3){a.x - b.x, a.y - b.y, a.z - b.z});
 }
 
-t_vec3	vec3_mul(t_vec3 a, float scalar)
+t_vec3	vec3_mul(t_vec3 a, double scalar)
 {
 	return ((t_vec3){a.x * scalar, a.y * scalar, a.z * scalar});
 }
@@ -33,7 +33,7 @@ t_vec3	vec3_normalize(t_vec3 vec)
 	return (vec3_mul(vec, 1 / vec3_mag(vec)));
 }
 
-float	vec3_mag(t_vec3 vec)
+double	vec3_mag(t_vec3 vec)
 {
-	return (sqrtf(vec.x * vec.x + vec.y * vec.y + vec.z * vec.z));
+	return (sqrt(vec.x * vec.x + vec.y * vec.y + vec.z * vec.z));
 }
